+++ lib/whois/record/scanners/whois.cnnic.cn.rb
@@ -27,20 +27,23 @@ module Whois
 
 
         tokenizer :scan_available do
-          if @input.scan(/^no matching record/)
+          if @input.match?(/^no matching record/)
             @ast["status:available"] = true
+            @input.scan_until(/\n/)
           end
         end
 
         tokenizer :scan_reserved do
-          if @input.scan(/^the domain you want to register is reserved/)
+          if @input.match?(/^the domain you want to register is reserved/)
             @ast["status:reserved"] = true
+            @input.scan_until(/\n/)
           end
         end
 
         tokenizer :scan_reserved_list do
           if @input.scan(/^Sorry, The domain you requested is in the reserved list/)
             @ast["status:reserved"] = true
+            @input.scan_until(/\n/)
           end
         end