+++ CHANGELOG.md
@@ -63,6 +63,8 @@
 
 - CHANGED: Updated whois.sx parser to the new response format.
 
+- CHANGED: Updated whois.cnnic.cn parser to the new response format.
+
 - FIXED: whois.domainregistry.ie should support status `Active - LOCKED`.
 
 - FIXED: whois.nic.uk fails to parse registrars without URL (GH-188).